+++ datafiles/mapper.py
@@ -13,11 +13,14 @@ from cached_property import cached_property
 
 from . import config, formats, hooks
 from .converters import Converter, List, map_type
-from .utils import display, recursive_update, write
-
-
-Trilean = Optional[bool]
-Missing = dataclasses._MISSING_TYPE
+from .utils import (
+    Missing,
+    Trilean,
+    display,
+    get_default_field_value,
+    recursive_update,
+    write,
+)
 
 
 class Mapper:
@@ -64,7 +67,7 @@ class Mapper:
         cls = self._instance.__class__
         try:
             root = Path(inspect.getfile(cls)).parent
-        except TypeError:  # pragma: no cover
+        except TypeError:
             level = log.DEBUG if '__main__' in str(cls) else log.WARNING
             log.log(level, f'Unable to determine module for {cls}')
             root = Path.cwd()
@@ -147,11 +150,11 @@ class Mapper:
                     value,
                     default_to_skip=Missing
                     if include_default_values
-                    else self._get_default_field_value(name),
+                    else get_default_field_value(self._instance, name),
                 )
 
             elif (
-                value == self._get_default_field_value(name)
+                value == get_default_field_value(self._instance, name)
                 and not include_default_values
             ):
                 log.debug(f"Skipped default value of {value!r} for {name!r} attribute")
@@ -178,9 +181,9 @@ class Mapper:
     def text(self, value: str):
         write(self.path, value.strip() + '\n')
 
-    def load(self, *, _log=True, _first=False) -> None:
+    def load(self, *, _log=True, _first_load=False) -> None:
         if self._root:
-            self._root.load(_log=_log, _first=_first)
+            self._root.load(_log=_log, _first_load=_first_load)
             return
 
         if self.path:
@@ -197,76 +200,48 @@ class Mapper:
 
             for name, value in data.items():
                 if name not in self.attrs and self.auto_attr:
-                    cls: Any = type(value)
-                    if issubclass(cls, list):
-                        cls.__origin__ = list
-
-                        if value:
-                            item_cls = type(value[0])
-                            for item in value:
-                                if not isinstance(item, item_cls):
-                                    log.warn(f'{name!r} list type cannot be inferred')
-                                    item_cls = Converter
-                                    break
-                        else:
-                            log.warn(f'{name!r} list type cannot be inferred')
-                            item_cls = Converter
-
-                        log.debug(f'Inferring {name!r} type: {cls} of {item_cls}')
-                        self.attrs[name] = map_type(cls, name=name, item_cls=item_cls)
-                    elif issubclass(cls, dict):
-                        cls.__origin__ = dict
-
-                        log.debug(f'Inferring {name!r} type: {cls}')
-                        self.attrs[name] = map_type(cls, name=name, item_cls=Converter)
-                    else:
-                        log.debug(f'Inferring {name!r} type: {cls}')
-                        self.attrs[name] = map_type(cls, name=name)
+                    self.attrs[name] = self._infer_attr(name, value)
 
             for name, converter in self.attrs.items():
-                log.debug(f"Converting '{name}' data with {converter}")
-
-                if getattr(converter, 'DATACLASS', None):
-                    self._set_dataclass_value(data, name, converter)
-                else:
-                    self._set_attribute_value(data, name, converter, _first)
+                self._set_value(self._instance, name, converter, data, _first_load)
 
             hooks.apply(self._instance, self)
 
         self.modified = False
 
-    def _set_dataclass_value(self, data, name, converter):
-        # TODO: Support nesting unlimited levels
-        # https://github.com/jacebrowning/datafiles/issues/22
-        nested_data = data.get(name)
-        if nested_data is None:
-            return
+    @staticmethod
+    def _infer_attr(name, value):
+        cls: Any = type(value)
+        if issubclass(cls, list):
+            cls.__origin__ = list
+            if value:
+                item_cls = type(value[0])
+                for item in value:
+                    if not isinstance(item, item_cls):
+                        log.warn(f'{name!r} list type cannot be inferred')
+                        item_cls = Converter
+                        break
+            else:
+                log.warn(f'{name!r} list type cannot be inferred')
+                item_cls = Converter
+            log.debug(f'Inferring {name!r} type: {cls} of {item_cls}')
+            return map_type(cls, name=name, item_cls=item_cls)
 
-        log.debug(f'Converting nested data to Python: {nested_data}')
+        if issubclass(cls, dict):
+            cls.__origin__ = dict
+            log.debug(f'Inferring {name!r} type: {cls}')
+            return map_type(cls, name=name, item_cls=Converter)
 
-        dataclass = getattr(self._instance, name)
-        if dataclass is None:
-            for field in dataclasses.fields(converter.DATACLASS):
-                if field.name not in nested_data:
-                    nested_data[field.name] = None
-            dataclass = converter.to_python_value(nested_data, target_object=dataclass)
+        log.debug(f'Inferring {name!r} type: {cls}')
+        return map_type(cls, name=name)
 
-        mapper = create_mapper(dataclass)
-        for name2, converter2 in mapper.attrs.items():
-            _value = nested_data.get(name2, mapper._get_default_field_value(name2))
-            value = converter2.to_python_value(
-                _value, target_object=getattr(dataclass, name2)
-            )
-            log.debug(f"'{name2}' as Python: {value!r}")
-            setattr(dataclass, name2, value)
+    @staticmethod
+    def _set_value(instance, name, converter, data, first_load):
+        log.debug(f"Converting '{name}' data with {converter}")
 
-        log.debug(f"Setting '{name}' value: {dataclass!r}")
-        setattr(self._instance, name, dataclass)
-
-    def _set_attribute_value(self, data, name, converter, first_load):
         file_value = data.get(name, Missing)
-        init_value = getattr(self._instance, name, Missing)
-        default_value = self._get_default_field_value(name)
+        init_value = getattr(instance, name, Missing)
+        default_value = get_default_field_value(instance, name)
 
         if first_load:
             log.debug(
@@ -291,21 +266,7 @@ class Mapper:
             value = converter.to_python_value(file_value, target_object=init_value)
 
         log.debug(f"Setting '{name}' value: {value!r}")
-        setattr(self._instance, name, value)
-
-    def _get_default_field_value(self, name):
-        for field in dataclasses.fields(self._instance):
-            if field.name == name:
-                if not isinstance(field.default, Missing):
-                    return field.default
-
-                if not isinstance(field.default_factory, Missing):  # type: ignore
-                    return field.default_factory()  # type: ignore
-
-                if not field.init and hasattr(self._instance, '__post_init__'):
-                    return getattr(self._instance, name)
-
-        return Missing
+        setattr(instance, name, value)
 
     def save(self, *, include_default_values: Trilean = None, _log=True) -> None:
         if self._root: