+++ datafiles/utils.py
@@ -1,16 +1,36 @@
 """Internal helper functions."""
 
+import dataclasses
 from contextlib import suppress
 from functools import lru_cache
 from pathlib import Path
 from pprint import pformat
 from shutil import get_terminal_size
-from typing import Dict, Union
+from typing import Dict, Optional, Union
 
 import log
 
 
+Trilean = Optional[bool]
+Missing = dataclasses._MISSING_TYPE
+
+
 cached = lru_cache()
+
+
+def get_default_field_value(instance, name):
+    for field in dataclasses.fields(instance):
+        if field.name == name:
+            if not isinstance(field.default, Missing):
+                return field.default
+
+            if not isinstance(field.default_factory, Missing):  # type: ignore
+                return field.default_factory()  # type: ignore
+
+            if not field.init and hasattr(instance, '__post_init__'):
+                return getattr(instance, name)
+
+    return Missing
 
 
 def prettify(value) -> str:
@@ -68,13 +88,12 @@ def write(filename_or_path: Union[str, Path], text: str) -> None:
         text = dedent(text)
 
     message = f'Writing file: {path}'
-    log.debug(message)
-    line = '=' * len(message)
+    line = '=' * (31 + len(message))
     if text:
         content = text.replace(' \n', '␠\n')
     else:
         content = '∅\n'
-    log.debug(line + '\n' + content + line)
+    log.debug(message + '\n' + line + '\n' + content + line)
 
     path.parent.mkdir(parents=True, exist_ok=True)
     path.write_text(text)
@@ -84,23 +103,22 @@ def read(filename: str) -> str:
     """Read text from a file with logging."""
     path = Path(filename).resolve()
     message = f'Reading file: {path}'
-    log.info(message)
-    line = '=' * len(message)
+    line = '=' * (31 + len(message))
     text = path.read_text()
     if text:
         content = text.replace(' \n', '␠\n')
     else:
         content = '∅\n'
-    log.debug(line + '\n' + content + line)
+    log.debug(message + '\n' + line + '\n' + content + line)
     return text
 
 
 def display(path: Path, data: Dict) -> None:
     """Display data read from a file."""
     message = f'Data from file: {path}'
-    log.debug(message)
-    line = '=' * len(message)
-    log.debug(line + '\n' + prettify(data) + '\n' + line)
+    line = '=' * (31 + len(message))
+    content = prettify(data)
+    log.debug(message + '\n' + line + '\n' + content + '\n' + line)
 
 
 def logbreak(message: str = "") -> None:
@@ -110,4 +128,4 @@ def logbreak(message: str = "") -> None:
         line = '-' * (width - len(message) - 1) + ' ' + message
     else:
         line = '-' * width
-    log.info(line)
+    log.critical(line)