+++ tests/test_file_inference.py
@@ -1,22 +1,19 @@
-import pytest
-
 from datafiles import auto
+from datafiles.utils import dedent, logbreak, read, write
 
 
-@pytest.mark.xfail(reason='https://bitbucket.org/ruamel/yaml/issues/73')
-def test_auto_with_sample_file(write, logbreak, expect, read, dedent):
-
+def test_auto_with_sample_file(expect):
     write(
         'tmp/sample.yml',
         """
 
         homogeneous_list:
-        - 1
-        - 2
+          - 1
+          - 2
 
         heterogeneous_list:
-        - 1
-        - 'abc'
+          - 1
+          - 'abc'
 
         empty_list: []
 
@@ -41,17 +38,43 @@ def test_auto_with_sample_file(write, logbreak, expect, read, dedent):
         """
 
         homogeneous_list:
-        - 1
-        - 2
-        - 3
-
+          - 1
+          - 2
+          - 3
         heterogeneous_list:
-        - 1
-        - abc
-        - 5.6
-
+          - 1
+          - 'abc'
+          - 5.6
         empty_list:
-        - 7.8
+          - 7.8
+        """
+    )
 
+
+def test_float_inference(expect):
+    write(
+        'tmp/sample.yml',
+        """
+        language: python
+        python:
+          - 3.7
+          - 3.8
+        """,
+    )
+
+    logbreak("Inferring object")
+    sample = auto('tmp/sample.yml')
+
+    logbreak("Updating attribute")
+    sample.python.append(4)
+
+    logbreak("Reading file")
+    expect(read('tmp/sample.yml')) == dedent(
+        """
+        language: python
+        python:
+          - 3.7
+          - 3.8
+          - 4.0
         """
     )